restructured project still sucks but it sucks less
