+++ requirements.txt
@@ -11,3 +11,5 @@ psutil>=5.8
 pyarrow>=3.0
 torch~=2.1.1
 joblib~=1.3.2
+
+imblearn~=0.0
